requirements.txt
requests
python-dotenv
replicate
langchain_core
langchain_community
langchain
transformers==4.35.2
gradio
torch
huggingface-hub==0.19.4
sentence-transformers
wikipedia-api
wikipedia
tavily-python
assemblyai
langchain[docarray]
pypdf
InstructorEmbedding
chromadb
tiktoken
django-cors-headers
django
django_rest_framework